# video/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Content Uploader Agent",
    "description": "AI-powered content analysis system with JWT authentication",
    "version": "1.0.0"
  },
  "paths": {
...
    "/download/{content_id}": {
      "get": {
        "tags": [
          "STEP 4: Content Access & Streaming"
        ],
        "summary": "Download",
        "description": "STEP 4B: Download content file",
        "operationId": "download_download__content_id__get",
        "parameters": [
          {
            "name": "content_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stream/{content_id}": {
      "get": {
        "tags": [
          "STEP 4: Content Access & Streaming"
        ],
        "summary": "Stream Video",
        "description": "STEP 4C: Stream video content with range support (authentication optional for analytics)",
        "operationId": "stream_video_stream__content_id__get",
        "parameters": [
          {
            "name": "content_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Id"
            }
          },
          {
            "name": "range_request",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Range Request"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/cdn/upload-url": {
      "get": {
        "tags": [
          "CDN & Pre-signed URLs"
        ],
        "summary": "Generate Upload Url",
        "description": "Generate pre-signed URL for direct frontend uploads",
        "operationId": "generate_upload_url_cdn_upload_url_get",
        "parameters": [
          {
            "name": "filename",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          },
          {
            "name": "content_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "application/octet-stream",
              "title": "Content Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cdn/download-url/{content_id}": {
      "get": {
        "tags": [
          "CDN & Pre-signed URLs"
        ],
        "summary": "Generate Download Url",
        "description": "Generate pre-signed URL for content downloads",
        "operationId": "generate_download_url_cdn_download_url__content_id__get",
        "parameters": [
          {
            "name": "content_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Id"
            }
          },
          {
            "name": "expires_in",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 3600,
              "title": "Expires In"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cdn/stream-url/{content_id}": {
      "get": {
        "tags": [
          "CDN & Pre-signed URLs"
        ],
        "summary": "Generate Stream Url",
        "description": "Generate optimized streaming URL with CDN support",
        "operationId": "generate_stream_url_cdn_stream_url__content_id__get",
        "parameters": [
          {
            "name": "content_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Id"
            }
          },
          {
            "name": "expires_in",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 7200,
              "title": "Expires In"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cdn/assets/{asset_type}/{filename}": {
      "get": {
        "tags": [
          "CDN & Pre-signed URLs"
        ],
        "summary": "Serve Static Assets",
        "description": "Serve static assets with CDN redirect",
        "operationId": "serve_static_assets_cdn_assets__asset_type___filename__get",
        "parameters": [
          {
            "name": "asset_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Type"
            }
          },
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cdn/purge-cache/{content_id}": {
      "get": {
        "tags": [
          "CDN & Pre-signed URLs"
        ],
        "summary": "Purge Cdn Cache",
        "description": "Purge CDN cache for specific content (admin only)",
        "operationId": "purge_cdn_cache_cdn_purge_cache__content_id__get",
        "parameters": [
          {
            "name": "content_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Id"
            }
          },
          {
            "name": "admin_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Admin Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}